Work in Progress - Timeline
timeline.html - output into grid-row rather than trying to alter a <ul> and put divs in with different classes in order to display the circles.

The hover is awful and is commented out.

Going to add dynamic div that displays weeks when a circle is clicked / hovered.

Updated my to-do list